Added apache auth to Ninja
By switching to the apache Auth driver in 
config/auth.php and setting the apache_login
to a valid relative path (for browser) to a
script where the following piece of PHP code 
is addaed:

  @session_name('ninjasession');
  session_start();
  $_SESSION['username'] = $_SERVER['REMOTE_USER'];
  header('Location: /ninja/');

it is possible leave the authentication over to 
apache but still use ninja as usual.
All user info will still be synched over from 
cgi.cfg and htpasswd.users as before.

It still has some small quirks left to be fixed:
 * There's a problem getting the intended
   URI requested by the user as it now will always
   redirect to tac/index. 
 * A user that hasn't any visible objects (not
   authorized for any hosts or services) might end
   up in an infinite loop when being redirected
   to default/show_login

+++ application/controllers/default.php
@@ -35,6 +35,16 @@ class Default_Controller extends Ninja_Controller  {
 
 	public function show_login()
 	{
+		if (Kohana::config('auth.driver') == 'apache') {
+			if (isset($_SESSION['username'])) {
+				Auth::instance()->driver->login($_SESSION['username'], false, false);
+				$this->apache_login();
+			} else {
+				header('location: ' . Kohana::config('auth.apache_login'));
+			}
+			exit;
+		}
+
 		$this->session->delete('auth_user');
 		$this->session->delete('nagios_access');
 		$this->session->delete('contact_id');
@@ -167,7 +177,29 @@ class Default_Controller extends Ninja_Controller  {
 	public function logout()
 	{
 		User_Model::logout_user();
+		if (Kohana::config('auth.driver') == 'apache') {
+			# unset some session variables
+			$this->session->delete('username');
+			$this->session->delete('auth_user');
+			$this->session->delete('nagios_access');
+			$this->session->delete('contact_id');
+			$this->template = $this->add_view('logged_out');
+			return;
+		}
 		url::redirect('default/');
+	}
+
+	/**
+	*	Finalize login using the apache driver
+	*/
+	public function apache_login()
+	{
+		if (empty($_SESSION['username']) || !Auth::instance()->logged_in()) {
+			die('Error!');
+		}
+
+		$user_data = ORM::factory('user')->where('username', $_SESSION['username'])->find();
+		User_Model::complete_login($user_data);
 	}
 
 	/**
